fixed typos

+++ arch-installer.sh
@@ -5,6 +5,7 @@ BLUE='\033[0;34m'
 CYAN='\033[0;36m'
 GREEN='\033[0;32m'
 WHITE='\033[0;37m'
+NC='\033[0m
 
 printf "${CYAN}                _           _           _        _ _ 
   __ _ _ __ ___| |__       (_)_ __  ___| |_ __ _| | | ___ _ __ 
@@ -32,12 +33,13 @@ ${WHITE}
 "
 
 format() {
+	printf "\n"
 	printf "\n${BLUE}[${WHITE}?${BLUE}] ext4 partition:${WHITE} $disk"
 	read -r -n1 ext4
 	printf "\n${BLUE}[${WHITE}?${BLUE}] swap partition:${WHITE} $disk"
 	read -r -n1 swap
 	
-	printf "\n\n"
+	printf "\n${CYAN}"
 	mkfs.ext4 $disk$ext4
 	mkswap $disk$swap
 	swapon $disk$swap
@@ -51,8 +53,8 @@ reflector -c Brazil -a 12 --sort rate --save /etc/pacman.d/mirrorlist >& /dev/nu
 
 while ! [ "$opt" = 'y' -o "$opt" = 'Y' -o "$opt" = 'n' -o "$opt" = 'N' ]
 do
-	printf "\n${BLUE}[${WHITE}?${BLUE}] is the disk pre-partitioned? (y/n) <-- "
-	read -rsn1 opt
+	printf "\n${BLUE}[${WHITE}?${BLUE}] is the disk pre-partitioned? ${GREEN}(y/n) ${CYAN}<-- "
+	read -r n1 opt
 done
 
 if [[ "$opt" == 'y' ]] || [[ "$opt" == 'Y' ]]
@@ -68,44 +70,69 @@ fi
 printf "${BLUE}[${WHITE}+${BLUE}] mounting ${GREEN}$disk$ext4 ${BLUE}to /mnt\n\n"
 mount $disk$ext4 /mnt
 
-printf "${BLUE}[${WHITE}+${BLUE}] starting base packages installation\n\n"
+printf "${BLUE}[${WHITE}+${BLUE}] starting base packages installation\n\n${CYAN}"
 pacstrap /mnt base linux linux-firmware vim intel-ucode xf86-video-intel
 
 printf "${BLUE}[${WHITE}+${BLUE}] generating ${GREEN}fstab\n\n"
 genfstab -U /mnt >> /mnt/etc/fstab
 
 printf "${BLUE}[${WHITE}+${BLUE}] writing the other half of the script to /mnt\n\n"
-cat << EOF > test.test.test
+cat << EOF > /mnt/arch-installer2.sh
 #!/bin/bash
 
+BLUE='\033[0;34m'
+CYAN='\033[0;36m'
+GREEN='\033[0;32m'
+WHITE='\033[0;37m'
+
+printf "\${BLUE}[\${WHITE}+\${BLUE}] adjusting timezone and locale\${CYAN}"
 ln -s /usr/share/zoneinfo/America/Sao_Paulo /etc/localtime
 hwclock --systohc
 timedatectl set-ntp true
 
 sed -i 's/#en_US.UTF-8/en_US.UTF-8/g' /etc/locale.gen
 locale-gen
-
 echo "LANG=en_US.UTF-8" > /etc/locale.conf
 echo "KEYMAP=br-abnt2" > /etc/vconsole.conf
+
+printf "\${BLUE}[\${WHITE}+\${BLUE}] setting hostname and configuring \${GREEN}/etc/hosts"
 echo "arch" > /etc/hostname
-echo "127.0.0.1		localhost arch" >> /etc/hosts
-echo "::1		localhost" >> /etc/hosts
+echo "127.0.0.1			localhost arch" >> /etc/hosts
+echo "::1			localhost" >> /etc/hosts
 
-read -r -p "${BLUE}[${WHITE}+${BLUE}] password: " pwd
+printf "\${BLUE}[\${WHITE}?\${BLUE}] adding user \${CYAN}blvckmetxl"
+printf "\${BLUE}[\${WHITE}?\${BLUE}] password:\${CYAN} "
+read -r pwd
 useradd -mG wheel,audio,video blvckmetxl
-echo "root:\$pwd" | chpasswd
 echo "blvckmetxl:\$pwd" | chpasswd
+sed -i 's/# %wheel ALL=(ALL) NOPASSWD: ALL/%wheel ALL=(ALL) NOPASSWD: ALL/g' /etc/sudoers
 
-pacman -S grub networkmanager zsh reflector wget git net-tools iwctl base-devel dosfstools mtools dialog wireless_tools wpa_supplicant linux-headers
+printf "\${BLUE}[\${WHITE}+\${BLUE}] install some more packages\${CYAN}"
+pacman -S base-devel dialog dosfstools git grub iwd linux-headers mtools net-tools networkmanager reflector terminator wireless_tools wget wpa_supplicant xorg zsh
 
+printf "\${BLUE}[\${WHITE}+\${BLUE}] installing and setting up grub\${CYAN}"
 grub-install $disk
 grub-mkconfig -o /boot/grub/grub.cfg
+
+clear
+printf "\${BLUE}[\${WHITE}+\${BLUE}] \${CYAN}exiting."
+exit
 EOF
 chmod +x /mnt/arch-installer2.sh
 
 printf "${BLUE}[${WHITE}+${BLUE}] chrooting into the fresh ${CYAN}arch ${BLUE}installation\n\n"
 arch-chroot /mnt ./arch-installer2.sh
 
-clear
-printf "${BLUE}[${WHITE}+${BLUE}]${CYAN} rebooting"
-reboot
+printf "${BLUE}[${WHITE}+${BLUE}]${CYAN} reboot? ${GREEN}(y/n) ${CYAN}<-- "
+while ! [ "$rb" = 'y' -o "$rb" = 'Y' -o "$rb" = 'n' -o "$rb" = 'N' ]
+do
+	read -r -n1 rb
+done
+
+if [[ "$rb" == 'y' ]] || [[ "$rb" == 'Y' ]]
+then
+	reboot
+else
+	printf "\n${NC}"
+	exit
+fi
